Edit

+++ src/py/flwr/common/differential_privacy.py
@@ -39,7 +39,7 @@ def get_norm(input_arrays: NDArrays) -> float:
 def add_gaussian_noise_inplace(input_arrays: NDArrays, std_dev: float) -> None:
     """Add Gaussian noise to each element of the input arrays."""
     for array in input_arrays:
-        noise = np.random.normal(0, std_dev, array.shape, dtype=array.dtype)
+        noise = np.random.normal(0, std_dev, array.shape).astype(array.dtype)
         array += noise
